Ændret wheiging int til double + lavet koden.

+++ 11_cdio_del1/src/view/IOperatorView.java
@@ -9,7 +9,7 @@ public interface IOperatorView {
 	public String getNewPassword();
 	public double getTara();
 	public double getBrutto();
-	public void showNetto(int Tara, int Brutto, int Netto);
+	public void showNetto(double Tara, double Brutto, double Netto);
 	public String getOprName();
 	public String getCPR();
 	public void showOpr (int oprId, String cpr, String name, String password);
